Create Phone Keypad Problem
Given a string containing digits from 2-9 inclusive, return all possible letter combinations that the number could represent. Return the answer in any order.

A mapping of digits to letters (just like on the telephone buttons) is given below. 
Note :1 does not map to any letters.

For example,
digits = "23" then the outputs will be in form of a vector of strings like -> ["ad","ae","af","bd","be","bf","cd","ce","cf"]